if saving to a store fails, move to the next available one
Instead of trying to save in the first store that can put_objet(),
get the list of all writable stores, and stop after the first success.

This also implicitely establish the interface of returning true if
put_object() succeeded, and false if it failed. Maybe throwing exceptions
would be better?